[docs] fix API ToC rendering for single Classes and Namespaces
