+++ docs/components/plugins/APIBox.tsx
@@ -1,13 +1,13 @@
 import { mergeClasses } from '@expo/styleguide';
 import type { PropsWithChildren } from 'react';
 
+import { APISectionPlatformTags } from '~/components/plugins/api/APISectionPlatformTags';
 import {
   H3Code,
   STYLES_APIBOX,
   STYLES_APIBOX_WRAPPER,
 } from '~/components/plugins/api/APISectionUtils';
 import { PlatformName } from '~/types/common';
-import { PlatformTags } from '~/ui/components/Tag';
 import { MONOSPACE } from '~/ui/components/Text';
 
 type APIBoxProps = PropsWithChildren<{
@@ -21,7 +21,7 @@ export const APIBox = ({ header, platforms, children, className }: APIBoxProps) 
     <div
       css={[STYLES_APIBOX, STYLES_APIBOX_WRAPPER]}
       className={mergeClasses(className, '!pb-4 last:[&>*]:!mb-1')}>
-      {platforms && <PlatformTags prefix="Only for:" platforms={platforms} />}
+      {platforms && <APISectionPlatformTags prefix="Only for:" userProvidedPlatforms={platforms} />}
       {header && (
         <H3Code tags={platforms}>
           <MONOSPACE weight="medium" className="wrap-anywhere">
