+++ docs/components/plugins/APISection.tsx
@@ -28,12 +28,6 @@ type Props = {
   forceVersion?: string;
   testRequire?: any;
   headersMapping?: Record<string, string>;
-
-  /**
-   * Whether to expose all classes props in the sidebar.
-   * @default true when the api has only one class, false otherwise.
-   */
-  exposeAllClassPropsInSidebar?: boolean;
 };
 
 const filterDataByKind = (
@@ -106,7 +100,6 @@ const renderAPI = (
     apiName,
     testRequire = undefined,
     headersMapping = {},
-    ...restProps
   }: Omit<Props, 'forceVersion'>
 ): JSX.Element => {
   try {
@@ -284,11 +277,7 @@ const renderAPI = (
         />
         <APISectionConstants data={constants} apiName={apiName} sdkVersion={sdkVersion} />
         <APISectionMethods data={hooks} header="Hooks" sdkVersion={sdkVersion} />
-        <APISectionClasses
-          data={classes}
-          sdkVersion={sdkVersion}
-          exposeAllClassPropsInSidebar={restProps.exposeAllClassPropsInSidebar}
-        />
+        <APISectionClasses data={classes} sdkVersion={sdkVersion} />
         {props && !componentsProps.length ? (
           <APISectionProps data={props} sdkVersion={sdkVersion} defaultProps={defaultProps} />
         ) : null}
