+++ docs/components/plugins/api/APISectionMethods.tsx
@@ -5,7 +5,6 @@ import { APIDataType } from '~/components/plugins/api/APIDataType';
 import {
   AccessorDefinitionData,
   MethodDefinitionData,
-  MethodParamData,
   PropData,
 } from '~/components/plugins/api/APIDataTypes';
 import { APISectionDeprecationNote } from '~/components/plugins/api/APISectionDeprecationNote';
@@ -18,7 +17,6 @@ import {
   STYLES_APIBOX,
   STYLES_APIBOX_NESTED,
   STYLES_NOT_EXPOSED_HEADER,
-  TypeDocKind,
   getH3CodeWithBaseNestingLevel,
   getTagData,
   getAllTagData,
@@ -151,63 +149,3 @@ const APISectionMethods = ({
   ) : null;
 
 export default APISectionMethods;
-
-export const APIMethod = ({
-  name,
-  sdkVersion,
-  comment,
-  returnTypeName,
-  isProperty = false,
-  isReturnTypeReference = false,
-  exposeInSidebar = false,
-  parameters = [],
-  platforms = [],
-}: {
-  exposeInSidebar?: boolean;
-  name: string;
-  sdkVersion: string;
-  comment: string;
-  returnTypeName: string;
-  isProperty: boolean;
-  isReturnTypeReference: boolean;
-  platforms: ('Android' | 'iOS' | 'Web')[];
-  parameters: {
-    name: string;
-    comment?: string;
-    typeName: string;
-    isReference?: boolean;
-  }[];
-}) => {
-  const parsedParameters = parameters.map(
-    param =>
-      ({
-        name: param.name,
-        type: { name: param.typeName, type: param.isReference ? 'reference' : 'literal' },
-
-        comment: {
-          summary: [{ kind: 'text', text: param.comment }],
-        },
-      }) as MethodParamData
-  );
-  return renderMethod(
-    {
-      name,
-      signatures: [
-        {
-          name,
-          parameters: parsedParameters,
-          comment: {
-            summary: [{ kind: 'text', text: comment }],
-            blockTags: platforms.map(text => ({
-              tag: 'platform',
-              content: [{ kind: 'text', text }],
-            })),
-          },
-          type: { name: returnTypeName, type: isReturnTypeReference ? 'reference' : 'literal' },
-        },
-      ],
-      kind: isProperty ? TypeDocKind.Property : TypeDocKind.Function,
-    },
-    { sdkVersion, exposeInSidebar }
-  );
-};
