+++ docs/pages/modules/module-api.mdx
@@ -6,7 +6,7 @@ sidebar_title: Module API
 
 import { APIBox } from '~/components/plugins/APIBox';
 import { CodeBlocksTable } from '~/components/plugins/CodeBlocksTable';
-import { APIMethod } from '~/components/plugins/api/APISectionMethods';
+import { APIMethod } from '~/components/plugins/api/APIMethod';
 import { FileTree } from '~/ui/components/FileTree';
 import { PlatformTags, StatusTag } from '~/ui/components/Tag';
 import { A, CALLOUT } from '~/ui/components/Text';
