+++ docs/pages/versions/unversioned/sdk/sqlite.mdx
@@ -3,7 +3,6 @@ title: SQLite
 description: A library that provides access to a database that can be queried through a SQLite API.
 sourceCodeUrl: 'https://github.com/expo/expo/tree/main/packages/expo-sqlite'
 packageName: 'expo-sqlite'
-maxHeadingDepth: 5
 iconUrl: '/static/images/packages/expo-sqlite.png'
 platforms: ['android', 'ios']
 ---
